+++ src/HayateOP.Extensions.HealthCheck/Microsoft/Extensions/DependencyInjection/ServiceCollectionExtensions.cs
@@ -6,7 +6,7 @@ namespace Microsoft.Extensions.DependencyInjection;
 public static class ServiceCollectionExtensions
 {
     public static IHayateServiceCollection RegisterHealthChecks<T>(this IHayateServiceCollection services)
-        where T : class, new()
+        where T : class
     {
         services.Services.AddHealthChecks().AddCheck<HayateOpHealthCheck<T>>($"HayateOpHealthCheck_{typeof(T).Name}");
         services.Services.AddScoped<HayateOpHealthCheck<T>>();